Separate launch BFF layer
