+++ apps/portal/src/app/api/launch/activate/route.ts
@@ -1,0 +1,4 @@
+import { activateLaunchRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const POST = activateLaunchRoute;
