+++ apps/portal/src/app/api/launch/app/route.ts
@@ -1,0 +1,4 @@
+import { launchAppRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const GET = launchAppRoute;
