+++ apps/portal/src/app/api/launch/check-repo/route.ts
@@ -1,0 +1,4 @@
+import { checkLaunchRepoRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const GET = checkLaunchRepoRoute;
