+++ apps/portal/src/app/api/launch/create/route.ts
@@ -1,0 +1,4 @@
+import { createLaunchRepoRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const POST = createLaunchRepoRoute;
