+++ apps/portal/src/app/api/launch/deploy/route.ts
@@ -1,0 +1,4 @@
+import { launchDeployRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const POST = launchDeployRoute(false);
