+++ apps/portal/src/app/api/launch/status/route.ts
@@ -1,0 +1,4 @@
+import { launchStatusRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const GET = launchStatusRoute;
