+++ apps/portal/src/app/api/launch/sync-installed/route.ts
@@ -1,0 +1,4 @@
+import { syncInstalledLaunchRoute } from "@portal/server/bff/launch/routes";
+
+export const runtime = "nodejs";
+export const POST = syncInstalledLaunchRoute;
