--- apps/portal/src/app/api/onboard/deploy/route.ts
@@ -1,3 +1,0 @@
-import { handleDeploy } from "@portal/lib/route-factory";
-export const runtime = "nodejs";
-export const POST = handleDeploy(false);
