--- apps/portal/src/app/api/onboard/dry-run/route.ts
@@ -1,3 +1,0 @@
-import { handleDeploy } from "@portal/lib/route-factory";
-export const runtime = "nodejs";
-export const POST = handleDeploy(true);
