+++ apps/portal/src/components/settings/onboarding/live-panel.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { CheckCircle2, ExternalLink } from "lucide-react";
-import { TEMPLATE_REPO_URL } from "@portal/lib/onboarding";
+import { TEMPLATE_REPO_URL } from "@portal/features/launch";
 
 /** Shared success state. `repo` is owner/name when known; falls back to the
  *  template so the clone story still renders during early wiring.
