+++ apps/portal/src/components/settings/onboarding/onboarding.tsx
@@ -5,17 +5,17 @@ import { CheckCircle2 } from "lucide-react";
 import { useAomiAuthAdapter } from "@aomi-labs/widget-lib";
 import {
   githubAppInstallUrl,
-  loadOnboarding,
-  saveOnboarding,
+  loadLaunch,
+  saveLaunch,
   readGithubRedirect,
   withPath,
   withProgress,
   withPendingInstall,
   GITHUB_REDIRECT_KEYS,
-  type OnboardingState,
-  type OnboardingPath,
-  type PathProgress,
-} from "@portal/lib/onboarding";
+  type LaunchState,
+  type LaunchPath,
+  type LaunchProgress,
+} from "@portal/features/launch";
 import { resolveDeployPlatform } from "@portal/lib/deploy-platform";
 import { Picker } from "./picker";
 import { OneshotWizard } from "./oneshot-wizard";
@@ -25,8 +25,8 @@ export function Onboarding() {
   const adapter = useAomiAuthAdapter();
   const actor = adapter.identity.address ?? undefined;
 
-  const [state, setState] = useState<OnboardingState>(() => {
-    const loaded = loadOnboarding();
+  const [state, setState] = useState<LaunchState>(() => {
+    const loaded = loadLaunch();
     if (typeof window !== "undefined") {
       const url = new URL(window.location.href);
       const urlDeployId = url.searchParams.get("deployment_id");
@@ -42,13 +42,13 @@ export function Onboarding() {
     return loaded;
   });
   const [installingPath, setInstallingPath] =
-    useState<OnboardingPath | null>(null);
+    useState<LaunchPath | null>(null);
   const [installError, setInstallError] = useState<string | null>(null);
   const [installSuccess, setInstallSuccess] = useState(false);
 
-  const update = useCallback((next: OnboardingState) => {
+  const update = useCallback((next: LaunchState) => {
     setState(next);
-    saveOnboarding(next);
+    saveLaunch(next);
   }, []);
 
   // --- hydrate the GitHub install redirect (runs once) ----------------------
@@ -58,21 +58,21 @@ export function Onboarding() {
     if (!redirect) {
       setInstallingPath(null);
       setInstallError(null);
-      const cur = loadOnboarding();
+      const cur = loadLaunch();
       if (cur.pendingInstall) {
         update(withPendingInstall(cur, null));
       }
       return;
     }
 
-    const cur = loadOnboarding();
+    const cur = loadLaunch();
     const matched = cur.pendingInstall?.path ?? cur.path;
     if (matched) {
       const next = withPendingInstall(
         withProgress(withPath(cur, matched), matched, {
           ...(redirect.repo ? { repo: redirect.repo } : {}),
           installationId: redirect.installationId,
-          installationStatus: redirect.onboard ?? undefined,
+          installationStatus: redirect.launchStatus ?? undefined,
         }),
         null,
       );
@@ -120,7 +120,7 @@ export function Onboarding() {
 
   // --- actions handed to the picker / wizards -------------------------------
   const choose = useCallback(
-    (path: OnboardingPath) => update(withPath(state, path)),
+    (path: LaunchPath) => update(withPath(state, path)),
     [state, update],
   );
 
@@ -130,13 +130,13 @@ export function Onboarding() {
   );
 
   const makePatch = useCallback(
-    (path: OnboardingPath) => (patch: Partial<PathProgress>) =>
+    (path: LaunchPath) => (patch: Partial<LaunchProgress>) =>
       update(withProgress(state, path, patch)),
     [state, update],
   );
 
   const makeOnReset = useCallback(
-    (path: OnboardingPath) => () => {
+    (path: LaunchPath) => () => {
       if (typeof window === "undefined") return;
       const url = new URL(window.location.href);
       url.searchParams.delete("deployment_id");
@@ -148,10 +148,10 @@ export function Onboarding() {
   );
 
   const makeBeginInstall = useCallback(
-    (path: OnboardingPath, mode: "install" | "authorize" = "install") =>
+    (path: LaunchPath, mode: "install" | "authorize" = "install") =>
       async () => {
         const next = withPendingInstall(withPath(state, path), { path });
-        saveOnboarding(next);
+        saveLaunch(next);
         setState(next);
         setInstallError(null);
         setInstallingPath(path);
